+++ examples/std/src/bin/net.rs
@@ -4,7 +4,7 @@ use clap::Parser;
 use embassy_executor::{Executor, Spawner};
 use embassy_net::StackStorage;
 use embassy_net::tcp::TcpSocket;
-use embassy_net::wire::{IpCidr, Ipv4Address};
+use embassy_net::wire::{IpCidr, Ipv4Addr};
 use embassy_net_tuntap::TunTapDevice;
 use embassy_time::Duration;
 use embedded_io_async::Write;
@@ -51,14 +51,14 @@ async fn main_task(spawner: Spawner) {
     // Choose between dhcp or static ip
     if opts.static_ip {
         iface
-            .add_ip_addr(IpCidr::new(Ipv4Address::new(192, 168, 69, 2).into(), 24))
+            .add_ip_addr(IpCidr::new(Ipv4Addr::new(192, 168, 69, 2).into(), 24))
             .unwrap();
         stack
             .routes()
-            .add_default_ipv4_route(Ipv4Address::new(192, 168, 69, 1), iface.handle())
+            .add_default_ipv4_route(Ipv4Addr::new(192, 168, 69, 1), iface.handle())
             .unwrap();
     } else {
-        iface.set_dhcpv4(Some(Default::default()));
+        iface.set_dhcpv4(Some(Default::default())).unwrap();
     }
 
     // Launch network task
@@ -71,9 +71,9 @@ async fn main_task(spawner: Spawner) {
 
     socket.set_timeout(Some(Duration::from_secs(10)));
 
-    let remote_endpoint = (Ipv4Address::new(192, 168, 69, 100), 8000);
-    info!("connecting to {:?}...", remote_endpoint);
-    let r = socket.connect(remote_endpoint).await;
+    let remote_addr = (Ipv4Addr::new(192, 168, 69, 100), 8000);
+    info!("connecting to {:?}...", remote_addr);
+    let r = socket.connect(remote_addr).await;
     if let Err(e) = r {
         warn!("connect error: {:?}", e);
         return;